requirements.txt
Django==2.1.2
pycryptodome==3.8.2
pytz==2019.1
